Update example_01
If SIRIUS is used to compute G-vectors then it also should be
used to compute the step function. The issue here is that step
function is the only place in Exciting where all G-vectors (ngrtot)
of the reciprocal lattice are used to compute the Fourier components
and then tranformed to real space. SIRIUS never generates all
G-vectors (up to ngrtot), so using the G-vectors of SIRIUS
(up to ngv) breaks the gencfun() of Exciting.